+++ proto/index.ts
@@ -1,0 +1,11 @@
+import * as SharedStreetsLinearReferencesProto from './linear_references'
+import * as SharedStreetsProto from './sharedstreets'
+import * as SharedStreetsSpeedsProto from './speeds'
+import * as SharedStreetsTemporalReferencesProto from './temporal_references'
+
+export {
+  SharedStreetsLinearReferencesProto,
+  SharedStreetsProto,
+  SharedStreetsSpeedsProto,
+  SharedStreetsTemporalReferencesProto,
+}